chore: another shuffle to decouple Tado and TadoX
