+++ src/index.ts
@@ -1,3 +1,3 @@
-export { Tado, TadoError } from "./tado";
+export { Tado } from "./tado";
 export { TadoX } from "./tadox";
 export * from "./types";
