+++ src/types/index.ts
@@ -40,6 +40,13 @@ import {
 
 export * from "./enums";
 
+/**
+ * TadoError extends the base Error class to represent errors specific to the Tado ecosystem.
+ * This custom error class can be used to differentiate between general errors
+ * and those that are Tado-specific, enabling more precise error handling and debugging.
+ */
+export class TadoError extends Error {}
+
 // utils
 export type DeepPartial<T> = T extends object
   ? {
